Do not write any files in dry mode #59

+++ bin/worker.execute.py
@@ -363,8 +363,9 @@ class WorkerExecution:
                     + "mkdir -p ~/.local/state/" + self.config.program_name + "/repos/public/net.splitcells.network\n"
                     + "cd ~/.local/state/" + self.config.program_name + "/repos/public/net.splitcells.network\n"
                     + Path(self.config.executable_path).read_text())
-            with open("./target/" + self.program_name, 'w') as executable_file:
-                executable_file.write(self.local_executable)
+            if not self.config.dry_run:
+                with open("./target/" + self.program_name, 'w') as executable_file:
+                    executable_file.write(self.local_executable)
             self.docker_file += "ADD ./" + self.program_name + " /root/program\n"
             self.docker_file += 'ENTRYPOINT /root/program'
         if self.config.class_for_execution is not None:
@@ -398,13 +399,14 @@ class WorkerExecution:
             self.docker_file = self.docker_file.replace('$ContainerSetupCommand', '\n')
         self.docker_file = self.docker_file.replace('${NAME_FOR_EXECUTION}', self.config.program_name)
         self.docker_file = self.docker_file.replace('${programName}', self.config.program_name)
-        file = 'target/Dockerfile-' + self.config.execution_name
-        if os.path.exists(file):
-            os.remove(file)
-        with open(file, 'w') as file_to_write:
-            file_to_write.write(self.docker_file)
-        with open('target/net.splitcells.network.worker.pom.xml', 'w') as pom_file_to_write:
-            pom_file_to_write.write(CONTAINER_POM)
+        if not self.config.dry_run:
+            file = 'target/Dockerfile-' + self.config.execution_name
+            if os.path.exists(file):
+                os.remove(file)
+            with open(file, 'w') as file_to_write:
+                file_to_write.write(self.docker_file)
+            with open('target/net.splitcells.network.worker.pom.xml', 'w') as pom_file_to_write:
+                pom_file_to_write.write(CONTAINER_POM)
         if self.config.only_execute_image:
             self.local_execution_script = PREPARE_EXECUTION_WITHOUT_BUILD_TEMPLATE
         else:
